+++ src/utils/Utils.java
@@ -10,7 +10,7 @@ public class Utils {
 
 	static int MAX_COORD_INTEGER = 16384;
 	
-	public static float[] getPosition(Entity e){
+	public static double[] getPosition(Entity e){
 		int cell_x = e.getProperty("m_cellX");
 		int cell_y = e.getProperty("m_cellY");
 		//System.out.println(e.toString());
@@ -18,9 +18,9 @@ public class Utils {
 	    int cellbits = e.getProperty("m_cellbits");
 	    		
 	    int cellwidth = 1 << cellbits;
-	    float[] pos = new float[2];
-	    pos[0] = ((cell_x * cellwidth) - MAX_COORD_INTEGER) + offset.x;
-	    pos[1] = ((cell_y * cellwidth) - MAX_COORD_INTEGER) + offset.y;
+	    double[] pos = new double[2];
+	    pos[0] = ((cell_x * cellwidth) - MAX_COORD_INTEGER) + (double)offset.x;
+	    pos[1] = ((cell_y * cellwidth) - MAX_COORD_INTEGER) + (double)offset.y;
 	    return pos;
 	}
 }
